+++ data/scripts/player/cosmicvaultcodex.lua
@@ -1,0 +1,15 @@
+package.path = package.path .. ";data/scripts/lib/?.lua"
+package.path = package.path .. ";data/scripts/?.lua"
+
+include("utility")
+
+function initialize()
+    if onClient() then
+        Player():registerCallback("onCosmicCodexGatherData", "onCosmicCodexGatherData")
+    end
+end
+
+function onCosmicCodexGatherData()
+    include("player/codex/infoCv")
+    infoCv_injectToCodex()
+end